+++ CHANGELOG.md
@@ -18,3 +18,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
  
  -  Temp fix for AHPS site status request failing
+ -  Fix for QCF value reset